requirements.txt
anthropic
feedparser
google-auth-oauthlib
google-api-python-client
requests
beautifulsoup4
pypdf
